Add unit test for solvestereopt
Adds a unit test, that should cover the entire chain of converting a Camera struct to a projection matrix and solving the stereo problem. Are you fine with the style of the code? I am not completely happy but lack a better idea. Maybe have the `@test` makro in the loop.