check visible points are in front of the camera
when computing the visibility of points in the `cameraproject` function,
also make sure that the points are in front of the camera. This can be
easily and cheaply done in this place, as the z position in camera local
coordinates is computed during projection.
The weird array syntax (`pt[3:3,:]`) was necessary for performance.
Whatever it does...