Bump required Julia to >=1.9
By requiring a higher Julia version, the package extension mechanism
will be available and can be used always (no need for Requires.jl
anymore). This makes the implementation easier.
The `get_plot_backend()` function is still contained to avoid method
overwrite warnings. Usually, a package extension does type piracy and
dispatches on a new type. We only have the demo functions, which cannot
dispatch on the type.
This now can be extended to different plotting backends (Makie for
example), which in turn could be selected in the `get_plot_backend()`
function.